Merge commit from fork
* Clean up events on EVM reversion in try-catch

* backport v0.3.x-specific file layout and helpers

* yihuang-patch adaptation for v0.3.x

Adds balance handler factory to un-refactored precompile code, adapting it for Distribution, Staking, Slashing, ICS20, Gov, and the now-deprecated Evidence. Skips the ERC20 as that one did not yet make use of the balance handler in v0.4.x.

* Add adapted test for v0.3.x

Added balance handler test, adapted for the v0.4.x precompile structure. Uses a decomposed version of the p.RunSetup function without the method checks.

Also, has to use Ginkgo tests with explicit state saving to prevent overwrites. Note that the events list here is much larger due to the structure of the factory contract caller.

* fix mocks2

* Add debug json

* add debug comments